fix(theme-builder): reorganize sources around core.ts
and drop dynamic-scheme.ts
